pyproject.toml
[project]
name = "florentine-abbot"
dynamic = ["version"]
description = "A suite of tools for archival scanning and digital workflow automation."
authors = [
    { name="Andrey Nalivayev", email="nalivayev@gmail.com" }
]
keywords = ["vuescan", "scanning", "automation", "archiving", "exif", "metadata"]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "Pillow>=10.0.0",
    "watchdog>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
face-core = [
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]
face-insightface = [
    "florentine-abbot[face-core]",
    "insightface>=0.7.3",
    "onnxruntime>=1.16.0",
]

[project.scripts]
scan-batcher = "scan_batcher.cli:main"
file-organizer = "file_organizer.cli:main"
preview-maker = "preview_maker.cli:main"
tile-cutter = "tile_cutter.cli:main"
face-recognizer = "face_recognizer.cli:main"
archive-keeper = "archive_keeper.cli:main"
florentine-web = "ui.web.app:main"
setup-runner = "setup_runner.cli:main"

[project.entry-points."florentine_abbot.tasks"]
preview-maker = "preview_maker.provider:MakerProvider"
tile-cutter = "tile_cutter.provider:CutterProvider"
face-recognizer = "face_recognizer.provider:RecognizerProvider"

[project.entry-points."face_recognizer.detectors"]
insightface = "face_recognizer.detectors.insightface.detector:InsightFaceDetector"

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "scan_batcher",
    "scan_batcher.workflows",
    "scan_batcher.workflows.vuescan",
    "file_organizer",
    "face_recognizer",
    "face_recognizer.detectors",
    "face_recognizer.detectors.insightface",
    "preview_maker",
    "tile_cutter",
    "archive_keeper",
    "common",
    "ui",
    "ui.web",
    "setup_runner",
]

[tool.setuptools.package-data]
"scan_batcher.workflows.vuescan" = ["*.ini"]
"common" = ["*.json"]
"file_organizer" = ["*.json"]
"face_recognizer" = ["*.json"]
"face_recognizer.detectors.insightface" = ["*.json"]
"preview_maker" = ["*.json"]
"tile_cutter" = ["*.json"]
"ui.web" = ["frontend/dist/**/*"]

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"

[tool.pytest.ini_options]
markers = [
    "manual: manual tests excluded from regular runs (run with: pytest -m manual)",
]
addopts = "-m 'not manual'"
testpaths = [
    "tests/common",
    "tests/file_organizer",
    "tests/integration",
    "tests/preview_maker",
    "tests/scan_batcher",
]
